+++ htable.h
@@ -17,7 +17,7 @@ extern htable htable_new(int capacity, hashing_t hash_type);
 extern int find_prime(int size);
 extern int is_prime(int candidate);
 extern void htable_delete(htable h);
-extern int htable_insert(htable h, const char *s, long docid);
+extern int htable_insert(htable h, const char *s, int docid);
 extern int htable_search(htable h, const char *s);
 extern void htable_print(htable h);
 extern int htable_save_to_disk(htable h, FILE*);
